Remove more code duplication
