+++ CMakeLists.txt
@@ -282,6 +282,7 @@ elseif (NINTENDO_SWITCH)
 elseif (${TARGET_PLATFORM} STREQUAL "android")
     set(PLATFORM_FILES
         ${PLATFORM_FILES}
+        ${PROJECT_SOURCE_DIR}/src/platform/android/android.c
         ${PROJECT_SOURCE_DIR}/src/platform/android/asset_handler.c
         ${PROJECT_SOURCE_DIR}/src/platform/android/jni.c
         ${PROJECT_SOURCE_DIR}/src/platform/SDL${SDL_VERSION}/android/android.c
