+++ src/platform/SDL2/keyboard_input.c
@@ -1,4 +1,4 @@
-#include "keyboard_input.h"
+#include "platform/keyboard_input.h"
 
 #include "game/cheats.h"
 #include "game/system.h"
@@ -7,6 +7,8 @@
 #include "input/keyboard.h"
 #include "input/mouse.h"
 #include "input/scroll.h"
+
+#include "SDL.h"
 
 static int is_alt_down(SDL_KeyboardEvent *event)
 {
@@ -226,14 +228,15 @@ static key_modifier_type get_modifier(int mod)
     return key_mod;
 }
 
-void platform_handle_key_down(SDL_KeyboardEvent *event)
+void platform_handle_key_down(void *event)
 {
+    SDL_KeyboardEvent *e = (SDL_KeyboardEvent *)event;
     // handle keyboard input keys
-    switch (event->keysym.sym) {
+    switch (e->keysym.sym) {
         case SDLK_RETURN:
         case SDLK_KP_ENTER:
             // only send enter if no modifier is also down
-            if ((event->keysym.mod & (KMOD_CTRL | KMOD_ALT | KMOD_GUI)) == 0) {
+            if ((e->keysym.mod & (KMOD_CTRL | KMOD_ALT | KMOD_GUI)) == 0) {
                 keyboard_return();
             }
             break;
@@ -266,7 +269,7 @@ void platform_handle_key_down(SDL_KeyboardEvent *event)
             break;
         case SDLK_AC_BACK:
 #ifndef __ANDROID__
-            event->keysym.scancode = SDL_SCANCODE_ESCAPE;
+            e->keysym.scancode = SDL_SCANCODE_ESCAPE;
             break;
 #else
         // Hack: since Android handles the right mouse button as a back button
@@ -277,14 +280,14 @@ void platform_handle_key_down(SDL_KeyboardEvent *event)
                 mouse_set_right_down(1);
                 return;
             } else {
-                event->keysym.scancode = SDL_SCANCODE_ESCAPE;
+                e->keysym.scancode = SDL_SCANCODE_ESCAPE;
             }
             break;
 #endif
     }
 
     // handle hotkeys
-    key_type key = get_key_from_scancode(event->keysym.scancode);
+    key_type key = get_key_from_scancode(e->keysym.scancode);
 
     if (keyboard_is_capturing()) {
         // Special event keys to handle when text input is active
@@ -306,12 +309,12 @@ void platform_handle_key_down(SDL_KeyboardEvent *event)
         }
     }
 
-    key_modifier_type mod = get_modifier(event->keysym.mod);
-    hotkey_key_pressed(key, mod, event->repeat);
+    key_modifier_type mod = get_modifier(e->keysym.mod);
+    hotkey_key_pressed(key, mod, e->repeat);
 
     // handle cheats: special case since they ARE layout dependent
-    if (!event->repeat && is_alt_down(event)) {
-        switch (event->keysym.sym) {
+    if (!e->repeat && is_alt_down(e)) {
+        switch (e->keysym.sym) {
             case SDLK_k:
                 game_cheat_activate();
                 break;
@@ -336,28 +339,31 @@ void platform_handle_key_down(SDL_KeyboardEvent *event)
     }
 }
 
-void platform_handle_key_up(SDL_KeyboardEvent *event)
+void platform_handle_key_up(void *event)
 {
+    SDL_KeyboardEvent *e = (SDL_KeyboardEvent *)event;
 #ifdef __ANDROID__
     // Right mouse button hack: read above for explanation
-    if ((event->keysym.sym == SDLK_ESCAPE || event->keysym.sym == SDLK_AC_BACK) && !mouse_get()->is_touch) {
+    if ((e->keysym.sym == SDLK_ESCAPE || e->keysym.sym == SDLK_AC_BACK) && !mouse_get()->is_touch) {
         mouse_set_right_down(0);
         return;
     }
 #endif
-    key_type key = get_key_from_scancode(event->keysym.scancode);
-    key_modifier_type mod = get_modifier(event->keysym.mod);
+    key_type key = get_key_from_scancode(e->keysym.scancode);
+    key_modifier_type mod = get_modifier(e->keysym.mod);
     hotkey_key_released(key, mod);
 }
 
-void platform_handle_editing_text(SDL_TextEditingEvent *event)
+void platform_handle_editing_text(void *event)
 {
-    keyboard_editing_text(event->text);
+    SDL_TextEditingEvent *e = (SDL_TextEditingEvent *)event;
+    keyboard_editing_text(e->text);
 }
 
-void platform_handle_text(SDL_TextInputEvent *event)
+void platform_handle_text(void *event)
 {
-    keyboard_text(event->text);
+    SDL_TextInputEvent *e = (SDL_TextInputEvent *)event;
+    keyboard_text(e->text);
 }
 
 key_type system_keyboard_key_for_symbol(const char *name)
