+++ src/platform/SDL3/joystick.c
@@ -1,8 +1,9 @@
-#include "joystick.h"
+#include "platform/joystick.h"
 
-#include "input/joystick.h"
 #include "input/touch.h"
 #include "platform/platform.h"
+
+#include <SDL3/SDL.h>
 
 #include <stdio.h>
 #include <string.h>
@@ -175,33 +176,6 @@ static gamepad_mapping gamepad_mappings[] = {
 
 #define NUM_GAMEPAD_MAPPINGS (sizeof(gamepad_mappings) / sizeof(gamepad_mapping))
 
-static int use_joystick(void)
-{
-#if defined(__vita__) || defined(__SWITCH__) || defined(__ANDROID__)
-    return 1;
-#else
-    return enabled;
-#endif
-}
-
-static joystick_hat_position convert_hat_position(int value)
-{
-    joystick_hat_position position = JOYSTICK_HAT_CENTERED;
-    if (value & SDL_HAT_UP) {
-        position |= JOYSTICK_HAT_UP;
-    }
-    if (value & SDL_HAT_DOWN) {
-        position |= JOYSTICK_HAT_DOWN;
-    }
-    if (value & SDL_HAT_LEFT) {
-        position |= JOYSTICK_HAT_LEFT;
-    }
-    if (value & SDL_HAT_RIGHT) {
-        position |= JOYSTICK_HAT_RIGHT;
-    }
-    return position;
-}
-
 static SDL_GamepadBinding *get_binding_for_element(SDL_GamepadBinding **bindings, int total_bindings,
     const gamepad_mapping_element *element)
 {
@@ -273,7 +247,7 @@ static void set_default_controller_mapping(joystick_model *model, SDL_Gamepad *g
                 case SDL_GAMEPAD_BINDTYPE_HAT:
                     action->element[j].type = JOYSTICK_ELEMENT_HAT;
                     action->element[j].id = binds[j]->input.hat.hat;
-                    action->element[j].position = convert_hat_position(binds[j]->input.hat.hat_mask);
+                    action->element[j].position = platform_joystick_convert_hat_position(binds[j]->input.hat.hat_mask);
                     break;
                 default:
                     break;
@@ -362,11 +336,20 @@ static void remove_joystick(SDL_JoystickID instance_id)
     joystick_remove(instance_id);
 }
 
+int platform_joystick_is_enabled(void)
+{
+#if defined(__vita__) || defined(__SWITCH__) || defined(__ANDROID__)
+    return 1;
+#else
+    return enabled;
+#endif
+}
+
 void platform_joystick_init(int force_enable)
 {
     enabled = force_enable;
 
-    if (!use_joystick()) {
+    if (!platform_joystick_is_enabled()) {
         return;
     }
     SDL_SetJoystickEventsEnabled(true);
@@ -378,46 +361,32 @@ void platform_joystick_init(int force_enable)
     SDL_free(instance_ids);
 }
 
-void platform_joystick_device_changed(SDL_JoystickID id, int is_connected)
+void platform_joystick_device_changed(long long id, int is_connected)
 {
-    if (!use_joystick()) {
+    if (!platform_joystick_is_enabled()) {
         return;
     }
     if (is_connected) {
-        add_joystick(id);
+        add_joystick((SDL_JoystickID) id);
     } else {
-        remove_joystick(id);
+        remove_joystick((SDL_JoystickID) id);
     }
 }
 
-void platform_joystick_handle_axis(SDL_JoyAxisEvent *event)
+joystick_hat_position platform_joystick_convert_hat_position(int value)
 {
-    if (!use_joystick()) {
-        return;
+    joystick_hat_position position = JOYSTICK_HAT_CENTERED;
+    if (value & SDL_HAT_UP) {
+        position |= JOYSTICK_HAT_UP;
     }
-    joystick_update_element(event->which, JOYSTICK_ELEMENT_AXIS, event->axis, event->value, 0);
-}
-
-void platform_joystick_handle_trackball(SDL_JoyBallEvent *event)
-{
-    if (!use_joystick()) {
-        return;
+    if (value & SDL_HAT_DOWN) {
+        position |= JOYSTICK_HAT_DOWN;
     }
-    joystick_update_element(event->which, JOYSTICK_ELEMENT_TRACKBALL, event->ball, event->xrel, event->yrel);
-}
-
-void platform_joystick_handle_hat(SDL_JoyHatEvent *event)
-{
-    if (!use_joystick()) {
-        return;
+    if (value & SDL_HAT_LEFT) {
+        position |= JOYSTICK_HAT_LEFT;
     }
-    joystick_update_element(event->which, JOYSTICK_ELEMENT_HAT, event->hat, convert_hat_position(event->value), 0);
-}
-
-void platform_joystick_handle_button(SDL_JoyButtonEvent *event, int is_down)
-{
-    if (!use_joystick()) {
-        return;
+    if (value & SDL_HAT_RIGHT) {
+        position |= JOYSTICK_HAT_RIGHT;
     }
-    joystick_update_element(event->which, JOYSTICK_ELEMENT_BUTTON, event->button, is_down, 0);
+    return position;
 }
