+++ src/platform/ios/ios.h
@@ -1,13 +1,12 @@
 #ifndef PLATFORM_IOS_H
 #define PLATFORM_IOS_H
 
-#include "SDL_platform.h"
-
-#ifdef __IPHONEOS__
+#if defined(TARGET_OS_IOS) && !defined(TARGET_OS_MACCATALYST)
 
 const char *ios_show_c3_path_dialog(int again);
 void c3_path_chosen(char *new_path);
 const char *ios_get_base_path(void);
 
-#endif // __IPHONEOS__
+#endif // defined(TARGET_OS_IOS) && !defined(TARGET_OS_MACCATALYST)
+
 #endif // PLATFORM_IOS_H
