+++ src/platform/touch.h
@@ -1,0 +1,8 @@
+#ifndef PLATFORM_TOUCH_H
+#define PLATFORM_TOUCH_H
+
+void platform_touch_start(void *event);
+void platform_touch_move(void *event);
+void platform_touch_end(void *event);
+
+#endif // PLATFORM_TOUCH_H
